spec: fix patching of files in subdirectories
With 23df04b ("dracut: create sub-directories for dracut modules") the
dracut modules were moved to subdirectories. This causes problems when
someone wants to include a patch to the spec file to change one of the
files in those subdirectories. Reason is that '%autosetup' in the spec
file calls 'patch' per default. 'patch' however, will strip all
directories when it is called without option -p. Which means that it
will search the file in the root directory and then fail because it
cannot find it. Thus add option -p1 to '%autosetup' which will be passed
on to 'patch'. Choose -p1 as that will work with the most common patch
creation tools like git and packit.

Signed-off-by: Philipp Rudo <prudo@redhat.com>