kdump-lib-initramfs: find mntpoint by btrfs subvol
When using btrfs dump target, we need to use subvol to find the real
mntpoint instead of the first output of findmnt, which is almost always
root subvol.

All functions that call this function are also updated.

For user configured btrfs dump target, we don't need the subvol and
it will always use the first subvol.

Signed-off-by: Lichen Liu <lichliu@redhat.com>