kdump.service: Replace ConditionKernelCommandLine with ExecCondition
Commit 0084806493d0 ("kdump.service: use ConditionKernelCommandLine=crashkernel")
added a condition based on the crashkernel kernel command line parameter to
control the start of the kdump service.

While ConditionKernelCommandLine=crashkernel works well for kdump, it causes
issues for fadump (specific to the PowerPC architecture), which also uses the
same service unit. Unlike kdump, crashkernel kernel command-line is not
mandatory for fadump.

Since ConditionKernelCommandLine doesn't support evaluating multiple kernel
command line parameters with dependencies between them, it has been replaced
with ExecCondition to resolve this limitation.

Now, if fadump is configured and the crashkernel parameter is NOT present in the
kernel command line, kdump service will start.

Fixes: #35
Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>