Set up crashkernel value for osbuild
Resolves: https://issues.redhat.com/browse/RHEL-63071

Currently the cloud image built by osbuild doesn't have crashkernel set
up. This happens because kdump-utils' posttrans scriptlet is executed
after kernel's one. As a result kdump.service hasn't been enabled when
kernel's posttrans scriptlet is executed. So when a new kernel is
installed and the kernel installation hook 92-crashkernel.install calls
kdumpctl reset-crashkernel-for-installed_kernel "$KERNEL_VERSION",
kdumpctl will refuse to set up the crashkernel because kdump.service
is yet to enabled.

This patch now let "systemd preset" kdump.service in the %post scriptlet
instead. And we expect users have now upgraded to kdump-utils so there
is no need for the solution introduced in commit 5fe098fe.

Note an agreement is reached between kdump-utils and osbuild in
https://bugzilla.redhat.com/show_bug.cgi?id=2024976 that kdump-utils
will be responsible for setting up the crashkernel kernel parameter. And
since kdump-utils will only set up crashkernel when kdump.service is
enabled, we need to make sure kdump.service is enabled in the systemd
preset.

Fixes: 5fe098fe ("Don't systemctl preset kdump when updating kexec-tools to kdump-utils")
Signed-off-by: Coiby Xu <coxu@redhat.com>