kdumpbase/kdump.sh: move ssh/scp calls to separate function
In dump_ssh the _ssh_opts need to be word split, which triggers
ShellChecks SC2086 warning. This warning is currently disabled for all
calls to ssh/scp. However in the ShellCheck wiki [1] the suggested solution
for POSIX compatible code is to make the call in a separate function.
This also makes disabling of SC2029 obsolete.

One big benefit of this solution is that now calls to ssh/scp can also
be made outside of dump_ssh without redefining _ssh_opts.

One small downside is that the options passed to ssh/scp must now be
maintained in each new function separately. Compared to the benefit
described above this is a small price to pay.

Note: Currently ssh/scp option -q,--quiet is used inconsistently. With
this commit it will always be set.

[1] https://www.shellcheck.net/wiki/SC2086

Signed-off-by: Philipp Rudo <prudo@redhat.com>