kdumpbase/kdump.sh: make use of global NEWROOT
The mount point for file system dumps is currently passed as an argument
to dump_fs. But there are only two possible values for the mount point.
Its either $NEWROOT for failure_action dump_to_rootfs or whatever the
user provided in kdump.conf. Thus instead of passing the mount point as
argument use a global variable and update it when parsing the config and
use that in dump_fs. Reuse $NEWROOT for this.

Signed-off-by: Philipp Rudo <prudo@redhat.com>