kdumpbase/kdump.sh: drop setting test status to 'failure' initially
When fetching the test status from the target a missing status file is
always considered a 'failure'. So there is no need to explicitly setting
the status to 'failure' in the initrd. This allows simplifying the code
a bit. For example we can now assume that the directory for $KDUMP_PATH
always exists (otherwise dump_{fs,ssh} would have returned with an
error).

Signed-off-by: Philipp Rudo <prudo@redhat.com>