kdumpctl/test: fix fetching of status for ssh target with ipv6
When trying to fetch the status from a ssh host that is specified via
its IPv6 address kdumpctl currently hangs. This is because scp requires
the IPv6 address to be enclosed in brackets, i.e. [], which is not the
case in kdump.conf. Fix this by copying what is done in kdump.sh in the
initrd.

While at it also add option -q to silence scp.

Fixes: 24e7622 ("Re-introduce vmcore creation notification to kdump")
Signed-off-by: Philipp Rudo <prudo@redhat.com>