fence_kdump: Shorten the interval between two message
Resolves: https://issues.redhat.com/browse/RHEL-46337

As man 8 fence_kdump_send:
-i, --interval=INTERVAL
Time to wait between sending a message. The value for INTERVAL must be greater than zero. (default: 10)

The interval 10 seconds are two large especially in the case that local
dumping goes fast. Suppose the following scenario:
	network is not ready
	fence_kdump_notify &
	network is ready
	local dumping finish and reboot within 10 seconds.
We will miss the chance to send out the fence dump messages.

Shorten the interval to one second to ease this issue.

Signed-off-by: Pingfan Liu <piliu@redhat.com>