tests: only wait 10s when checking vmcore
Currently, the checking vmcore test fails but the test still passes and
is treated as an error. This is because the test wait 5 minutes for the
vmcore to be ready and exceed the maximum test time 5m

        execute task #4: Check vmcore on server
        how: tmt
        progress:                                                   warn: Test /Check vmcore/tests/check_vmcore/ssh failed, stopping execution.
        progress:

        summary: 4 tests executed
    report
        how: display
        summary: 3 tests passed and 1 error

So simply wait for 10s instead. Another reason is the vmcore checking
test will always be running after the kernel panic test on the client so
there is no need to wait for so long.

Signed-off-by: Coiby Xu <coxu@redhat.com>