kdumpctl: improve do_estimate by calculating initramfs 3 times.
A memory peak in kdump occurs when the initramfs is being unpacked. At this
moment, both the packed and unpacked initramfs coexist in memory.

Previously, I used some methods to calculate the size of the unpacked
initrd. However, I found that this was unnecessary. With the help of
Coiby [1], I found that the main problem was that when unpacking
initramfs, the actual available memory size for storing rootfs was only
half of the totalmem at that time.

This phenomenon can be seen by adding a df command to the dracut module
99squash-lib/init-squash.sh for debugging.

```
[ 0.540867] Trying to unpack rootfs image as initramfs...
[ 0.541228] INITRAMFS: Memory before decompression: TotalRAM=920368 FreeRAM=901000 KB
[ 0.543602] ACPI: bus type thunderbolt registered
[ 0.544006] clocksource: tsc: mask: 0xffffffffffffffff max_cycles: 0x2b2c8ec87c7, max_idle_ns: 440795278598 ns
[ 0.555883] INITRAMFS: Memory after decompression: TotalRAM=920368 FreeRAM=853348 KB
[ 0.560789] Freeing initrd memory: 47112K
...
[1.306324] Run /init as init process
Filesystem Size Used Avail Use% Mounted on
rootfs 450M 47M 404M 11% /
devtmpfs 450M 0 450M 0% /dev
tmpfs 476M 0 476M 0% /run
```

After initramfs unpacked, the origin cpio archive is released, given that the
size of initramfs is usually larger than 32M, the freed space is enough to
offset the runtime reserve memory, So I removed it here.

[1] https://github.com/rhkdump/kdump-utils/pull/78#issuecomment-2896488781

Signed-off-by: Lichen Liu <lichliu@redhat.com>