Forward core collector's stdout and stderr to /dev/console
The default core collector, makedumpfile, has functionality to print the
ETA for dump capture. However, due to a limitation/bug in systemd,
executing the kdump.sh script via a systemd service does not print the
ETA on the console, even though StandardError is set to journal+console.

On machines with high memory in the terabyte range, processing the dump
takes time. Since makedumpfile’s ETA is not printed on the console while
collecting the dump, there have been a few instances where system
administrators rebooted the system during dump capture, assuming the
system was stuck.

Until the systemd service issue [1] is fixed, forward the core
collector’s stdout and stderr to /dev/console manually. This will print
the ETA for dump capture and let the user know that dump capture is
ongoing.

[1] https://github.com/systemd/systemd/issues/37994

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>