kdumpctl: warn when reset crashkernel for low memory systems
The function of reset-crashkernel is to set crashkernel to the default
value  of  kdump-utils. However,  our  current  default value does not
reserve  memory  when  the  operating system memory is less than 2G or
fadump  is  enabled and the memory is less than 4G. Add a warn message
to  let  the user know that no memory will be reserved for crashkernel
due to insufficient system memory.

This  patch  does  not  prevent kdumpctl from setting the crashkernel,
because even if the crashkernel cannot be reserved, it is harmless and
could be useful for systems that are used as vm templates.

Signed-off-by: Lichen Liu <lichliu@redhat.com>