kdumpctl: clean up {backup,restore}_default_initrd
The sha1 algorithm used is no longer considered secure. In addition the
shaXsum commands have a --check option that reads checksums stored in a
file and verifies them. So there is no need for the home grown
verification in restore_default_initrd.

While at it refractor the two functions slightly to increase their
readability and add additional debug messages.

Signed-off-by: Philipp Rudo <prudo@redhat.com>