kdumpctl: add comments to different initrd variables
kdumpctl has multiple variables for different initrds it is using. From
the variable names it is not always clear what the difference of those
initrds are. Thus add some comments to describe the differences and what
the initrds are used for.

While at it rename INITRD_CHECKSUM_LOCATION to DEFAULT_INITRD_CHECKSUM
so it aligns better with the naming convention.

Signed-off-by: Philipp Rudo <prudo@redhat.com>