powerpc: consider CPU count while calculating crashkernel value
The next patch in the series adds more CPUs to the capture kernel,
which increases the memory requirement for the capture kernel.
Experiments show that powerpc needs 1 MB of additional memory for every
CPU added.

Therefore, while calculating the crashkernel size, make sure to include
an additional 1 MB for every CPU configured in the capture kernel.

The changes are implemented in such a way that if the user changes the
nr_cpus value in the kdump configuration, the script will adapt
accordingly.

If other architectures decide to enable more CPUs in the kdump kernel,
then defining the _pcpu_area variable in
kdump_get_arch_recommend_crashkernel()/kdump-lib.sh function will be
sufficient for this solution.

Signed-off-by: Sourabh Jain <sourabhjain@linux.ibm.com>