kdump-lib: always disable systemd-gpt-auto-generator
Resolves: KEX-330

The systemd-gpt-auto-generator automatically discovers and mounts
various partitions, e.g. the root partition, based on the GPT partition
type GUIDs. So far anaconda hasn't set the GUIDs during an installation
and users relying on it, e.g. for UKIs where the root= parameter cannot
be passed on the kernel command line, had to set them manually. This
seems to have changed and causes the kdump kernel to hang. In addition
we know which disk/partition shall be used as dump target and add it
explicitly to the kdump initrd. So it is save to always disable the
gpt-auto-generator for the kdump kernel.

Signed-off-by: Philipp Rudo <prudo@redhat.com>