kdump-lib: Skip adding reserved memory for small-memory machine
This increased "crashkernel=" raises the risk of OOM conditions in the
production kernel, particularly on small-memory machines that typically
lack such hardware features (Mellanox network cards, SMMU). So we
skip that case on aarch64. For other arches, the above factors are not
considered, so ignore them for the time being.

Signed-off-by: Pingfan Liu <piliu@redhat.com>