Wait for LUKS configfs API to be ready
Partially resolves: https://github.com/rhkdump/kdump-utils/issues/109

On Fedora 42, LUKS-encrypted volumes fail to be unlocked in the crash
kernel because somehow the volume keys are yet to be available. Wait for
the API to be ready before triggering restoring volume keys so that
later the keys will be available for cryptsetup to unlock the volumes.

Fixes: d9677e17 ("Support dumping to a LUKS-encrypted target")

Signed-off-by: Coiby Xu <coxu@redhat.com>