kdump.sh: Skip num-threads when -E and -F option is present
Resolves: https://issues.redhat.com/browse/RHEL-75537

Configure "makedumpfile -E -d 31" in kdump.conf, and panic the system, kdump will fail with the error as:

[ 41.891856] kdump.sh[724]: --num-threads cannot used with ELF format.
[ 41.897104] kdump.sh[724]: Commandline parameter is invalid.
[ 41.897290] kdump.sh[724]: Try `makedumpfile --help' for more information.
[ 41.897435] kdump.sh[724]: makedumpfile Failed.
[ 41.898804] kdump[726]:
saving vmcore failed, exitcode:1

Skip --num-threads when -E option is given.

As for the -F option, it is meaningless to sequentially access linear
addresses with multiple threads Skip --num-threads too in that case.

Signed-off-by: Pingfan Liu <piliu@redhat.com>