sysconfig: use initramfs_options to reduce memory usage
Starting with Linux 6.18, Fedora 42, RHEL 10.2, and RHEL 9.8, the
kernel supports the `initramfs_options` boot parameter.

Sets `initramfs_options=size=90%` can increase the size of the tmpfs
used for initramfs decompression from the default 50% to 90% of memory.

This change greatly reduces memory pressure during decompression,
allowing for a reduction in the memory reserved for the crashkernel.

Signed-off-by: Lichen Liu <lichliu@redhat.com>